requirements.txt
claude-agent-sdk==0.1.17
litellm[proxy]
python-dotenv
anthropic